api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = "^2.0.30"
pymysql = "^1.1.1"
ruff = "^0.4.8"
pytest = "^8.2.2"
alembic = "^1.13.1"
aiomysql = "^0.2.0"
firebase-admin = "^6.5.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.7"
aiosqlite = "^0.20.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
